requirements.txt
Flask
requests
Pillow
gunicorn